% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adan.analyze.R
\name{adan.analyze}
\alias{adan.analyze}
\title{Perform Prediction Reliability and Applicablity Domain assasemnt}
\usage{
adan.analyze(adan.model)
}
\arguments{
\item{adan.model}{A adan model built with adan.test.}
}
\description{
Perform Prediction Reliability and Applicablity Domain assasemnt
}

