% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adan.build.R
\name{adan.build}
\alias{adan.build}
\title{Perform Prediction Reliability and Applicablity Domain assasemnt}
\usage{
adan.build(train.md, train.a, train.p, scale.md = F, ncomp = NULL)
}
\arguments{
\item{train.md}{A matrix with molecular descriptors.}

\item{train.a}{A vector with activities of train data.}

\item{train.p}{A vector with predictions of trian data.}

\item{scale.md}{A boolean if molecular descriptors should be unit variance scaled.}
}
\value{
An adan model.
}
\description{
Perform Prediction Reliability and Applicablity Domain assasemnt
}

