% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adan.test.R
\name{adan.test}
\alias{adan.test}
\title{Perform Prediction Reliability and Applicablity Domain assasemnt}
\usage{
adan.test(adan.model, query.md, query.p)
}
\arguments{
\item{adan.model}{A adan model built with adan.test.}

\item{query.md}{A matrix with molecular descriptors with query data.}

\item{query.p}{A vector with predictions for query data.}
}
\value{
A list with applicability domain assesment and confidence interval estimation.
}
\description{
Perform Prediction Reliability and Applicablity Domain assasemnt
}

